+++ stock-service/src/main/java/com/example/stockService/domain/dto/Order.java
@@ -1,4 +1,4 @@
-package com.example.stockService;
+package com.example.stockService.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
