+++ stock-service/src/main/java/com/example/stockService/interfaces/eventhandlers/OderConsumer.java
@@ -1,5 +1,6 @@
-package com.example.stockService;
+package com.example.stockService.interfaces.eventhandlers;
 
+import com.example.stockService.domain.dto.Order;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 
@@ -8,7 +9,7 @@ import java.util.function.Consumer;
 @Service
 public class OderConsumer {
     @Bean
-    public Consumer<com.example.stockService.Order> input() {
+    public Consumer<Order> input() {
         System.out.println("Input Order");
         return order -> {
             String name = order.getName();
